// frontend/src/app/core/generated/circabc/model/logSearchResult.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LogSearchResult { 
    logDate: string;
    activityDescription: string;
    serviceDescription?: string;
    userId: string;
    information: string;
    path: string;
    success: boolean;
}